+++ express-zod-api/src/result-handler.ts
@@ -6,12 +6,7 @@ import {
   defaultStatusCodes,
   NormalizedResponse,
 } from "./api-response";
-import {
-  FlatObject,
-  isObject,
-  isSchema,
-  pullResponseExamples,
-} from "./common-helpers";
+import { FlatObject, isObject, isSchema } from "./common-helpers";
 import { contentTypes } from "./content-type";
 import { IOSchema } from "./io-schema";
 import { ActualLogger } from "./logger-helpers";
@@ -21,6 +16,7 @@ import {
   getPublicErrorMessage,
   logServerError,
   normalize,
+  pullResponseExamples,
   ResultSchema,
 } from "./result-helpers";
