+++ express-zod-api/tests/common-helpers.spec.ts
@@ -6,7 +6,6 @@ import {
   getMessageFromError,
   makeCleanId,
   ensureError,
-  pullResponseExamples,
   getRoutePathParams,
 } from "../src/common-helpers";
 import { z } from "zod/v4";
@@ -196,24 +195,6 @@ describe("Common Helpers", () => {
       expect(
         getMessageFromError(new Error("something went wrong")),
       ).toMatchSnapshot();
-    });
-  });
-
-  describe("pullResponseExamples()", () => {
-    test("handles multiple examples per property", () => {
-      const schema = z.object({
-        a: z.string().example("one").example("two").example("three"),
-        b: z.number().example(1).example(2),
-        c: z.boolean().example(false),
-      });
-      expect(pullResponseExamples(schema)).toEqual([
-        { a: "one", b: 1, c: false },
-        { a: "one", b: 2, c: false },
-        { a: "two", b: 1, c: false },
-        { a: "two", b: 2, c: false },
-        { a: "three", b: 1, c: false },
-        { a: "three", b: 2, c: false },
-      ]);
     });
   });
